pyproject.toml
[tool.poetry]
name = "gcp-phone-location"
version = "0.1.0"
description = ""
authors = ["VitaminB16 <artemiy.nosov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.2"
gcp-pal = "^1.0.29"
google-cloud-firestore = "^2.16.0"
google-cloud-logging = "^3.10.0"
selenium = "^4.21.0"


[tool.poetry.group.dev.dependencies]
gcsfs = "^2024.5.0"
google-cloud-functions = "^1.16.3"
google-cloud-scheduler = "^2.13.3"
pandas = "^2.2.2"
ipykernel = "^6.29.4"
beautifulsoup4 = "^4.12.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
